refactor: reorganize client into feature directories
Move files into feature-based directories:
- features/visualization/ — AlbumGrid, SongCard, CoverOverlay, ProgressBar,
  FlippableTile, VisualizationLayout, useAlbumGrid, useDominantColor
- features/playback/ — SpotifyContext, spotifyReducer, useNowPlayingPoller,
  usePlaybackSSE, api, mapToNowPlaying
- features/auth/ — UserContext, UserAvatar, UserMenu, SpotifyLoginButton
- components/ — shared: LoadingScreen, FullscreenButton
- hooks/ — shared: useWindowSize
