+++ packages/client/src/features/auth/UserAvatar.tsx
