+++ packages/client/src/features/playback/api.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import type { UserProfile, PlaybackState, SpotifyAlbum } from '../types';
+import type { UserProfile, PlaybackState, SpotifyAlbum } from '../../types';
 
 export async function getAuthUser(): Promise<UserProfile | null> {
   const response = await axios.get<{ user: UserProfile | null }>('/api/auth/user');
