+++ packages/client/src/features/visualization/AlbumImage.tsx
