+++ packages/client/src/features/visualization/CoverOverlay.css
