+++ packages/client/src/features/visualization/CoverOverlay.tsx
