+++ packages/client/src/features/visualization/ProgressBar.tsx
