+++ packages/client/src/pages/VisualizationContent.tsx
@@ -1,16 +1,16 @@
 import { useState, useCallback, useEffect } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
-import { useUser } from '../contexts/UserContext';
-import { useSpotify } from '../contexts/SpotifyContext';
-import useNowPlayingPoller from '../hooks/useNowPlayingPoller';
+import { useUser } from '../features/auth/UserContext';
+import { useSpotify } from '../features/playback/SpotifyContext';
+import useNowPlayingPoller from '../features/playback/useNowPlayingPoller';
 import useWindowSize from '../hooks/useWindowSize';
-import useAlbumGrid from '../hooks/useAlbumGrid';
-import useDominantColor from '../hooks/useDominantColor';
+import useAlbumGrid from '../features/visualization/useAlbumGrid';
+import useDominantColor from '../features/visualization/useDominantColor';
 import LoadingScreen from '../components/LoadingScreen';
-import VisualizationLayout from '../components/VisualizationLayout';
-import UserAvatar from '../components/UserAvatar';
-import UserMenu from '../components/UserMenu';
+import VisualizationLayout from '../features/visualization/VisualizationLayout';
+import UserAvatar from '../features/auth/UserAvatar';
+import UserMenu from '../features/auth/UserMenu';
 import FullscreenButton from '../components/FullscreenButton';
 import CastButton from '../cast/sender/CastButton';
 import useCastSession from '../cast/sender/useCastSession';
